+++ exts/base.py
@@ -26,7 +26,7 @@ class QrmHelpCommand(commands.HelpCommand):
         super().__init__(command_attrs={
             "help": "Shows help about qrm or a command",
             "aliases": ["h"],
-            # TODO "category": cmn.BoltCats.INFO
+            "category": cmn.BoltCats.INFO
             })
         self.verify_checks = True
         self.context: commands.Context
@@ -140,7 +140,7 @@ class BaseCog(commands.Cog):
             self.bot_invite = (f"https://discordapp.com/oauth2/authorize?client_id={self.bot.user.id}"
                                f"&scope=bot&permissions={opt.invite_perms}")
 
-    @commands.command(name="info", aliases=["about"])  # TODO , category=cmn.BoltCats.INFO)
+    @commands.command(name="info", aliases=["about"], category=cmn.BoltCats.INFO)
     async def _info(self, ctx: commands.Context):
         """Shows info about qrm."""
         embed = cmn.embed_factory(ctx)
@@ -158,7 +158,7 @@ class BaseCog(commands.Cog):
         embed.set_thumbnail(url=str(self.bot.user.avatar_url))
         await ctx.send(embed=embed)
 
-    @commands.command(name="ping", aliases=["beep"])  # TODO , category=cmn.BoltCats.INFO)
+    @commands.command(name="ping", aliases=["beep"], category=cmn.BoltCats.INFO)
     async def _ping(self, ctx: commands.Context):
         """Shows the current latency to the discord endpoint."""
         embed = cmn.embed_factory(ctx)
@@ -171,7 +171,7 @@ class BaseCog(commands.Cog):
         embed.description = f"Current ping is {self.bot.latency*1000:.1f} ms"
         await ctx.send(content, embed=embed)
 
-    @commands.command(name="changelog", aliases=["clog"])  # TODO , category=cmn.BoltCats.INFO)
+    @commands.command(name="changelog", aliases=["clog"], category=cmn.BoltCats.INFO)
     async def _changelog(self, ctx: commands.Context, version: str = "latest"):
         """Shows what has changed in a bot version. Defaults to the latest version."""
         embed = cmn.embed_factory(ctx)
@@ -207,7 +207,7 @@ class BaseCog(commands.Cog):
 
         await ctx.send(embed=embed)
 
-    @commands.command(name="issue")  # TODO , category=cmn.BoltCats.INFO)
+    @commands.command(name="issue", category=cmn.BoltCats.INFO)
     async def _issue(self, ctx: commands.Context):
         """Shows how to create a bug report or feature request about the bot."""
         embed = cmn.embed_factory(ctx)
@@ -219,7 +219,7 @@ class BaseCog(commands.Cog):
                             [miaowware/qrm-resources](https://github.com/miaowware/qrm-resources/issues)."""
         await ctx.send(embed=embed)
 
-    @commands.command(name="donate", aliases=["tip"])  # TODO , category=cmn.BoltCats.INFO)
+    @commands.command(name="donate", aliases=["tip"], category=cmn.BoltCats.INFO)
     async def _donate(self, ctx: commands.Context):
         """Shows ways to help support development of the bot via donations."""
         embed = cmn.embed_factory(ctx)
@@ -230,7 +230,7 @@ class BaseCog(commands.Cog):
             embed.add_field(name=title, value=url, inline=False)
         await ctx.send(embed=embed)
 
-    @commands.command(name="invite", enabled=opt.enable_invite_cmd)  # TODO , category=cmn.BoltCats.INFO)
+    @commands.command(name="invite", enabled=opt.enable_invite_cmd, category=cmn.BoltCats.INFO)
     async def _invite(self, ctx: commands.Context):
         """Generates a link to invite the bot to a server."""
         if not (await self.bot.application_info()).bot_public:
